add initial version of openlapse sketch
Cut down version of the sonoff-tasmota firmware for iTead, Sonoff,
Wemos, NodeMCU hardware by Theo Arends

As this is a timelapse controller, it only needs a bit of a
web interface and a way to control a motor, so all the sensor
and MQTT code has been removed. Still a work-in-progress, but
it you can run a timelapse, control the motor and camera (if
you're using the openlapse baseboard).

It creates a WiFi HotSpot called 'openlapse', so associat to it,
then just browse to openlapse.local, and you should see the web
interface.

Signed-off-by: David Hunt <dave@davidhunt.ie>